requirements.txt
h5py>=3.0
albumentations>=1.4.18
cmapy>=0.6.6
matplotlib>=3.8.0
numpy>=1.21.0,<2.0.0
onnx>=1.17.0
onnxruntime>=1.21.0
opencv-contrib-python>=4.10.0.84
opencv-python>=4.10.0.84
opencv-python-headless>=4.11.0.86
Pillow>=11.3.0
scikit-learn
pytorch-model-summary>=0.1.2
thop>=0.1.1.post2209072238
torch>=2.7.0
torchvision>=0.22.0
tensorboard>=2.20.0
tensorboard-data-server>=0.7.2